working initialization script for SRAM1 in CM7 startup

+++ CM7/Core/App/sin_gen.h
@@ -107,4 +107,4 @@ struct sin_gen_envelop_generator
 void sin_gen_process(void);
 void sin_gen_init(void);
 void sin_gen_set_play(bool flag, uint8_t key_number);
-void sin_gen_set_envelop_generator(double sustain_level, uint32_t attack_time, uint32_t decay_time, uint32_t release_time);
+void sin_gen_set_envelop_generator(uint8_t sustain_level, uint32_t attack_time, uint32_t decay_time, uint32_t release_time);
